src/pyproject.toml
[tool.poetry]
name = "tictactoe"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
prometheus-fastapi-instrumentator = "^7.0.0"
flake8 = "^7.1.1"
orjson = "^3.10.7"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.34"}
flake8-docstrings = "^1.7.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
redis = "^5.0.8"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
ruff = "^0.6.3"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
